+++ kairo-rest/endpoint/src/main/kotlin/kairo/rest/KairoType.kt
@@ -3,7 +3,7 @@ package kairo.rest
 import io.ktor.util.reflect.TypeInfo
 import kairo.reflect.KairoType
 
-public fun <T : Any> KairoType<T>.toKtor(): TypeInfo =
+public fun <T> KairoType<T>.toKtor(): TypeInfo =
   TypeInfo(
     type = kotlinClass,
     kotlinType = kotlinType,